Move tests to tests.py
