+++ residents/tests.py
@@ -1,5 +1,9 @@
-from django.core.exceptions import ValidationError
+from io import StringIO
+from django.core.management import call_command
 from django.test import TestCase
+from django.core.management.base import CommandError
+
+from django.core.exceptions import ValidationError
 from unittest.mock import MagicMock, patch
 
 from .models import Residency, Resident
@@ -103,3 +107,105 @@ class TestResidentActivityLevel(TestCase):
         self.resident.on_hiatus = True
         expected = {"color": "info", "text": "On hiatus"}
         self.assertEqual(self.resident.activity_level, expected)
+
+
+class MakeResidencyTest(TestCase):
+    def test_residency_count(self):
+        out = StringIO()
+        call_command("make_fake_residencies", 3, stdout=out)
+        residency_count = Residency.objects.count()
+        self.assertEqual(residency_count, 3)
+        self.assertIn(
+            "Created " + str(residency_count) + " fake residencies.",
+            out.getvalue(),
+        )
+
+    def test_residency_no_residencies(self):
+        out = StringIO()
+        call_command("make_fake_residencies", 0, stdout=out)
+        residency_count = Residency.objects.count()
+        self.assertEqual(residency_count, 0)
+        self.assertIn(
+            "Created " + str(residency_count) + " fake residencies.",
+            out.getvalue(),
+        )
+
+    def test_residency_negative_number(self):
+        out = StringIO()
+        call_command("make_fake_residencies", -30, stdout=out)
+        self.assertIn("Invalid n. Please try again.", out.getvalue())
+
+    def test_residency_invalid_string_number(self):
+        out = StringIO()
+        try:
+            call_command("make_fake_residencies", "two", stdout=out)
+        except CommandError:
+            residency_count = Residency.objects.count()
+            self.assertEqual(residency_count, 0)
+
+    def test_residency_invalid_string(self):
+        try:
+            out = StringIO()
+            call_command("make_fake_residencies", "hello", stdout=out)
+        except CommandError:
+            residency_count = Residency.objects.count()
+            self.assertEqual(residency_count, 0)
+
+    def test_residency_number_string_mix(self):
+        try:
+            out = StringIO()
+            call_command("make_fake_residencies", "hello", stdout=out)
+        except CommandError:
+            residency_count = Residency.objects.count()
+            self.assertEqual(residency_count, 0)
+
+
+class MakeResidentTest(TestCase):
+    def test_resident_count(self):
+        out = StringIO()
+        call_command("make_fake_residents", 3, stdout=out)
+        resident_count = Resident.objects.count()
+        self.assertEqual(resident_count, 3)
+        self.assertIn(
+            "Created " + str(resident_count) + " fake residents.",
+            out.getvalue(),
+        )
+
+    def test_resident_no_residents(self):
+        out = StringIO()
+        call_command("make_fake_residents", 0, stdout=out)
+        resident_count = Resident.objects.count()
+        self.assertEqual(resident_count, 0)
+        self.assertIn(
+            "Created " + str(resident_count) + " fake residents.",
+            out.getvalue(),
+        )
+
+    def test_resident_negative_number(self):
+        out = StringIO()
+        call_command("make_fake_residents", -30, stdout=out)
+        self.assertIn("Invalid n. Please try again.", out.getvalue())
+
+    def test_resident_invalid_string_number(self):
+        out = StringIO()
+        try:
+            call_command("make_fake_residents", "two", stdout=out)
+        except CommandError:
+            resident_count = Resident.objects.count()
+            self.assertEqual(resident_count, 0)
+
+    def test_resident_invalid_string(self):
+        try:
+            out = StringIO()
+            call_command("make_fake_residents", "hello", stdout=out)
+        except CommandError:
+            resident_count = Resident.objects.count()
+            self.assertEqual(resident_count, 0)
+
+    def test_resident_number_string_mix(self):
+        try:
+            out = StringIO()
+            call_command("make_fake_residents", "hello", stdout=out)
+        except CommandError:
+            resident_count = Resident.objects.count()
+            self.assertEqual(resident_count, 0)
